Extract PriorityForm from ChangeValidityForm
