+++ ui/src/components/forms/common/index.ts
@@ -1,5 +1,4 @@
 export * from './ChangeValidityForm';
-export * from './customZodSchemas';
 export * from './DateRange';
 export * from './EnumDropdown';
 export * from './FormColumn';
@@ -7,4 +6,6 @@ export * from './FormRow';
 export * from './InputElement';
 export * from './InputField';
 export * from './InputLabel';
+export * from './PriorityForm';
 export * from './ValidationErrorList';
+export * from './customZodSchemas';
